Move tooltip to utility file
